Added Alternative Voting Framework
Currently only 2 voting systems exist, and this is Plurality and Reverse-Plurality, or "LOSER TAKES ALL". LTA shows who would have won in an election if we selected the candidate with the least votes. Its purely for interesting visualisations. Isn't it interesting how Tories are never unpopular enough to win a seat in this system? Or how Greens are the least popular recognised party, only narrowly bested by Independants, and this is all despite them receiving a good chunk of the nation-wide vote?