ElectoralMap now builds data based on groups
ElectoralMap now uses a different kind of data template. Initially it cared too much about if what was given to it was an individual constituency. However, it was recognisable that the functions could be rewritten to be agnostic of that detail, and instead only asks for the number of MPs elected and the voting data. It also now returns an array of winners. How this will be represented on the map is still a difficult question.

Four ideas are
-Just colour the constituency based on the biggest winner (this isn't ideal because it doesn't visualise the data without comprimse)
-Colour the constituency borders with the overall winner, and map the constituency colours by associating each winner with a constituency, based on what constituency contributed the most 1st votes for that winner. The borders would show the overall winner for the group, but you can see all of the winners on the map, even if inaccuracies appear from the assumption that the 1st choices contributed the most to that victory (sometimes the 2nd, 3rd, etc choice votes is what decides a candidate in STV), but for Plurality Multimember constituencies, it works great!
3. Stripe the constituencies with the colours of every winner. For every win from a party, the more weight they have in the stripes.

The 2nd is my preferred option, but it might look bad, in which case, 3 is the best choice.